requirements.txt
# Core
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# LLM integration
ollama>=0.1.7
httpx>=0.24.0

# NLP / Metrics
jiwer>=3.0.3                   # WER computation
sentence-transformers>=2.2.2   # Semantic similarity embeddings
transformers>=4.35.0
scikit-learn>=1.3.0

# Audio
soundfile>=0.12.1
librosa>=0.10.0

# Config & utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0                   # Pretty terminal output

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Dev
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
